+++ Doodle_Jump_Multiplayer/src/me/eluch/libgdx/DoJuMu/network/packets/PingDatas.java
@@ -28,7 +28,7 @@ public abstract class PingDatas {
 			ping = packet.readLong();
 			player = players.getPlayerByID(id);
 			if (player != null) {
-				player.setPing(ping);
+				player.setPing((int) ping);
 			}
 		}
 	}
